% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_get_overlap_region.R
\name{ensembl_get_overlap_region}
\alias{ensembl_get_overlap_region}
\title{Request features for a region}
\usage{
ensembl_get_overlap_region(regions, build = "GRCh38")
}
\arguments{
\item{regions}{A character vector of genomic regions e.g.
""1:55516888-55516888"}

\item{build}{Either 'GRCh38' (default) or 'GRCh37'.}
}
\value{
A tibble.
}
\description{
Retrieves features (e.g. genes, transcripts, variants and more) that overlap
a given region.
}
\examples{
 ensembl_get_overlap_region(regions = "1:55516888-55516888", build = "GRCh37")
}
