% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchGeneName.R
\name{OpenTargetFunc}
\alias{OpenTargetFunc}
\title{Nearest gene for given variant}
\usage{
OpenTargetFunc(rsids)
}
\arguments{
\item{rsids}{Character vector of rsids}
}
\value{
A data frame
}
\description{
Nearest gene for given variant
}
\examples{
OpenTargetFunc("rs12345")
}
