% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentargets.R
\name{ot_knowndrugs}
\alias{ot_knowndrugs}
\title{Get drugs known to target a given EnsemblId product}
\usage{
ot_knowndrugs(ensgIds, size = 1000)
}
\arguments{
\item{ensgIds}{Character vector of EnsemblIDs}

\item{size}{Maximum number of rows to return per EnsemblID queried}
}
\value{
A tibble
}
\description{
Get drugs known to target a given EnsemblId product
}
\examples{

ot_knowndrugs("ENSG00000157764")

}
