% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentargets.R
\name{ot_target_tractability}
\alias{ot_target_tractability}
\title{Get tractability data for EnsemblIDs}
\usage{
ot_target_tractability(ensgIds)
}
\arguments{
\item{ensgIds}{Character vector of EnsemblIDs}
}
\value{
A tibble
}
\description{
Returns a tibble of available \href{https://platform-docs.opentargets.org/target/tractability}{tractability data} for one or
more EnsemblIDs.
}
\details{
Each queried EnsemblID will return 4 rows of data, one for
each \href{https://platform-docs.opentargets.org/target/tractability#assessments}{assessment modality}:
\itemize{
\item 'SM' = Small molecule
\item 'AB' = Antibody
\item 'PR' = PROTAC
\item 'OC' = Other modalities (\emph{see \href{https://platform.opentargets.org/target/ENSG00000169083}{example} on OpenTargets web browser})
}
}
\examples{
# get tractability data for 2 targets
result <- ot_target_tractability(c("ENSG00000157764", "ENSG00000122482"))

# see if these are 'druggable', as per the pipeline by Finan et al
result |>
  dplyr::filter(modality == "SM") |>
  dplyr::select(ensgId, druggable_family)
}
