% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchGeneName.R
\name{OpenTargetFunc}
\alias{OpenTargetFunc}
\title{Nearest gene for given variant}
\usage{
OpenTargetFunc(query_rsID)
}
\arguments{
\item{query_rsID}{rsid}
}
\value{
dataframe
}
\description{
Nearest gene for given variant
}
\examples{
OpenTargetFunc("rs12345")
}
